requirement.txt
gradio
google-generativeai
 